Fix producer test names.
